Split indexes impl into different files
