+++ packages/storage-plus/src/lib.rs
@@ -8,9 +8,11 @@ mod item;
 mod iter_helpers;
 mod keys;
 mod map;
+mod multi_index;
 mod path;
 mod prefix;
 mod snapshot;
+mod unique_index;
 
 pub use endian::Endian;
 #[cfg(feature = "iterator")]
@@ -18,15 +20,15 @@ pub use indexed_map::{IndexList, IndexedMap};
 #[cfg(feature = "iterator")]
 pub use indexed_snapshot::IndexedSnapshotMap;
 #[cfg(feature = "iterator")]
-pub use indexes::{
-    index_string, index_string_tuple, index_triple, index_tuple, Index, MultiIndex, UniqueIndex,
-};
+pub use indexes::{index_string, index_string_tuple, index_triple, index_tuple, Index};
 pub use item::Item;
 pub use keys::{I128Key, I16Key, I32Key, I64Key, I8Key};
 pub use keys::{Prefixer, PrimaryKey, U128Key, U16Key, U32Key, U64Key, U8Key};
 pub use map::Map;
+pub use multi_index::MultiIndex;
 pub use path::Path;
 #[cfg(feature = "iterator")]
 pub use prefix::{range_with_prefix, Bound, Prefix};
 #[cfg(feature = "iterator")]
 pub use snapshot::{SnapshotItem, SnapshotMap, Strategy};
+pub use unique_index::UniqueIndex;
